Add rt_common module
